python/requirements.txt
python-dotenv
pandas
dune-client 